requirements.txt
streamlit
google-generativeai
ddgs>=2.0.0
python-dotenv
Pillow
